Canvas text: add 'async_layout' method
This method calls for asynchronous layout operation on the textblock
object.
There is some preparation work that may be requires - it will still be
done on the mainloop (due to requirement to load fonts through evas
font cache); afterwards the rest of the visual layout is offloaded to a
thread.
The mainloop and the thread negotiate when accessing the common
resources (i.e. the text layout). There is not wait, though - the work
is deferred.

@feature