Canvas text async: add async layout functionality
This adds Efl.Canvas.Text_Async with the 'async_layout' method.
The 'async layout' method is similar to 'size_formatted_get', but done
outside of the mainloop. When a call is made to this method, a thread
is created (after some preparation like updating the logical text
items), and the visual layout is offloaded to that thread.

The object also provides events to tackle problems such as attempts to
query or manipulate the text layout from the mainloop, after a thread
was created and hasn't yet complete its work. Higher-level widgets that
choose to use this object should register to these events and re-apply
such operations when the text layout is completed and the respective
callback of completion is called.

The usage of events is intentional, to prevent any need to override
methods of this object by higher-level widgets.

@feature