Ui text: add ability to set item factories
Added API to set an item factory object.
This is similar to the previous item providers (that worked with
callbacks).

You instantiate a factory object and set it on the Ui.Text object.
Each factory implements the "create" method from
Efl.Canvas.Text.Item_Factory.

This also includes 3 public factories (Image, Emoticon and Fallback), and one
The fallback provider can be disabled via Ui.Text API, for users that
do not want to preserve the fallback behavior.

See the added "Ui.text Item Factory" test in elementary_test for an
example of usage.

@feature